@@ -16,32 +16,35 @@
     "src/",
     "static/",
     "config.toml",
-    "svelte.config.cjs"
+    "svelte.config.js"
   ],
   "scripts": {
     "build": "svelte-kit build",
     "dev": "svelte-kit dev",
     "lint:css": "stylelint src/**/*.svelte",
     "lint:js": "eslint src --ext .js,.svelte",
-    "start": "svelte-kit start"
+    "preview": "svelte-kit preview"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "^1.0.0-next.4",
-    "@sveltejs/kit": "^1.0.0-next.72",
-    "rollup": "^2.45.1",
-    "svelte": "^3.37.0",
-    "vite": "^2.1.5",
+    "@sveltejs/adapter-static": "^1.0.0-next.13",
+    "@sveltejs/kit": "^1.0.0-next.129",
+    "rollup": "^2.53.2",
+    "svelte": "^3.38.3",
+    "vite": "^2.4.2",
     "vite-plugin-toml": "^0.3.1"
   },
   "optionalDependencies": {
-    "eslint": "^7.24.0",
-    "eslint-plugin-svelte3": "^3.1.2",
-    "stylelint": "^13.12.0"
+    "eslint": "^7.30.0",
+    "eslint-plugin-svelte3": "^3.2.0",
+    "stylelint": "^13.13.1"
   },
   "resolutions": {
+    "browserslist": "^4.16.6",
     "ini": "^2.0.0",
     "lodash": "^4.17.21",
-    "source-map-resolve": "^0.6.0"
+    "postcss": "^8.3.5",
+    "source-map-resolve": "^0.6.0",
+    "trim-newlines": "^4.0.2"
   },
   "eslintConfig": {
     "root": true,
@@ -330,14 +333,14 @@
       },
       {
         "files": [
-          "svelte.config.cjs"
+          "svelte.config.js"
         ],
         "env": {
           "node": true,
           "browser": false
         },
         "parserOptions": {
-          "sourceType": "script"
+          "sourceType": "module"
         }
       }
     ]
